requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.23.0
arxiv>=1.4.7
streamlit>=1.20.0
